Add check goal (fixes #22)
